requirements.txt
openai>=1.6.1,<2.0.0
python-dotenv==0.21.1
langchain==0.0.267
langchain-openai==0.0.2
pytest==7.4.3
chromadb==0.4.22
tiktoken==0.5.2 